requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg[binary]==3.2.1
pydantic==2.7.4
python-dotenv==1.0.1